pyproject.toml
[project]
name = "monoclaw"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai==2.30.0",
    "pydantic==2.12.5",
    "pydantic-settings==2.13.1",
    "pyyaml==6.0.3",
    "httpx==0.28.1",
    "croniter==6.2.2",
    "markdownify==1.2.2",
    "json-repair==0.58.7",
    "websockets==14.2",
]

[dependency-groups]
dev = ["black==26.3.1", "flake8==7.3.0", "flake8-pyproject==1.2.4", "pydocstyle==6.3.0", "mypy==1.20.0", "types-PyYAML==6.0.12.20250915"]

[tool.mypy]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["llm", "scheduler"]
disable_error_code = ["import-untyped"]

[tool.pydocstyle]
add-ignore = ["D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107"]

[tool.black]
line-length = 121
target-version = ["py311"]

[tool.flake8]
max-line-length = 121
extend-ignore = ["E203"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["src/*.py"]
